section commit for ERMS-6506
pre-final commit, tests of procedures outstanding

+++ grails-app/controllers/de/laser/ajax/AjaxController.groovy
@@ -25,7 +25,6 @@ import de.laser.remote.GlobalRecordSource
 import de.laser.storage.PropertyStore
 import de.laser.storage.RDConstants
 import de.laser.storage.RDStore
-import de.laser.survey.SurveyConfig
 import de.laser.survey.SurveyOrg
 import de.laser.survey.SurveyResult
 import de.laser.utils.CodeUtils
@@ -1311,47 +1310,63 @@ class AjaxController {
                             m.setProperty(prop, owner.getProperty(prop))
                             m.save()
                         }
-                    }
-                    else {
-                        AuditConfig.removeConfig(owner, prop)
 
-                        if (! params.keep) {
-                            members.each { m ->
-                                if(m[prop] instanceof Boolean)
-                                    m.setProperty(prop, false)
-                                else {
-                                    if(m[prop] instanceof RefdataValue) {
-                                        if(m[prop].owner.desc == RDConstants.Y_N_U)
-                                            m.setProperty(prop, RDStore.YNU_UNKNOWN)
-                                        else m.setProperty(prop, null)
+                        if(prop == 'holdingSelection' && members && owner[prop] == RDStore.SUBSCRIPTION_HOLDING_PARTIAL) {
+                            if(!subscriptionService.checkThreadRunning('PackageUnlink_'+owner.id)) {
+                                executorService.execute({
+                                    String threadName = 'PackageUnlink_' + owner.id
+                                    Thread.currentThread().setName(threadName)
+                                    //synchronise data
+                                    Set<Package> subPackages = Package.executeQuery('select sp.pkg from SubscriptionPackage sp where sp.subscription = :owner', [owner: owner])
+                                    subPackages.each { Package pkg ->
+                                        batchQueryService.clearIssueEntitlements([sub: members, pkg_id: pkg.id])
                                     }
-                                    else
-                                        m.setProperty(prop, null)
-                                }
-                                m.save()
+                                })
                             }
                         }
+                    }
+                    else {
+                        //accidental prevention
+                        if(owner[prop] != RDStore.SUBSCRIPTION_HOLDING_ENTIRE) {
+                            AuditConfig.removeConfig(owner, prop)
 
-                        // delete pending changes
-                        // e.g. PendingChange.changeDoc = {changeTarget, changeType, changeDoc:{OID,  event}}
-                        members.each { m ->
-                            List<PendingChange> openPD = PendingChange.executeQuery("select pc from PendingChange as pc where pc.status is null and pc.costItem is null and pc.oid = :objectID",
-                                    [objectID: "${m.class.name}:${m.id}"])
+                            if (! params.keep) {
+                                members.each { m ->
+                                    if(m[prop] instanceof Boolean)
+                                        m.setProperty(prop, false)
+                                    else {
+                                        if(m[prop] instanceof RefdataValue) {
+                                            if(m[prop].owner.desc == RDConstants.Y_N_U)
+                                                m.setProperty(prop, RDStore.YNU_UNKNOWN)
+                                            else m.setProperty(prop, null)
+                                        }
+                                        else
+                                            m.setProperty(prop, null)
+                                    }
+                                    m.save()
+                                }
+                            }
 
-                            openPD?.each { pc ->
-                                def payload = JSON.parse(pc?.payload)
-                                if (payload && payload?.changeDoc) {
-                                    def eventObj = genericOIDService.resolveOID(payload.changeDoc?.OID)
-                                    def eventProp = payload.changeDoc?.prop
-                                    if (eventObj?.id == owner?.id && eventProp.equalsIgnoreCase(prop)) {
-                                        pc.delete()
+                            // delete pending changes
+                            // e.g. PendingChange.changeDoc = {changeTarget, changeType, changeDoc:{OID,  event}}
+                            members.each { m ->
+                                List<PendingChange> openPD = PendingChange.executeQuery("select pc from PendingChange as pc where pc.status is null and pc.costItem is null and pc.oid = :objectID",
+                                        [objectID: "${m.class.name}:${m.id}"])
+
+                                openPD?.each { pc ->
+                                    def payload = JSON.parse(pc?.payload)
+                                    if (payload && payload?.changeDoc) {
+                                        def eventObj = genericOIDService.resolveOID(payload.changeDoc?.OID)
+                                        def eventProp = payload.changeDoc?.prop
+                                        if (eventObj?.id == owner?.id && eventProp.equalsIgnoreCase(prop)) {
+                                            pc.delete()
+                                        }
                                     }
                                 }
                             }
                         }
                     }
                 }
-
             }
         }
         if(Boolean.valueOf(params.returnSuccessAsJSON))
